Minor fixes to control and serialconnect
- Added close to serial connect in GUIManager
- Reworked controlSpeed

+++ src/ui/RobotSpeed.java
@@ -4,8 +4,8 @@ package ui;
  * Created by Lance on 5/2/2016.
  */
 public class RobotSpeed {
-    short speedMod = 50;
+    double speedMod = 50;
 
-    public void setSpeed(short d) { speedMod = d; }
-    public short getSpeed() {return speedMod;}
+    public void setSpeed(double d) { speedMod = d; }
+    public double getSpeed() {return speedMod;}
 }
